Updates for 1.14 and some fixes
This adds support for 1.14 and it's new block with some minor fixed, including a semi-work-around for MC-145113

NOTE: Due to MC-145113 ( https://bugs.mojang.com/browse/MC-145113 ), the traditional method of running multiple command from one command block utilizing falling block entites with passengers of command block minecarts doesn't currently work as the minecarts land inside the falling block instead of on top of it.  I'm currently putting armor stands with 0 health between every entity, as this raises the passengers riding the armor stand up an additional block.  However it takes a few ticks for the armor stand to land, die, and the have it's passengers fall and land again. This causes a single command block to take more than a second to process. This is problematic for the flying machine that activates all the command blocks, as it now activates the command blocks faster than they can be processed, leasing to a build up of entities, severe game lag (seconds per frame) or possible game crashes, even in single player.

That is why this is in a separate branch until this issue is fixed or a better work-around can be found.